ai/detector/requirements.txt
# openvino and its dependencies
openvino-dev==2023.0.0
openvino-telemetry==2022.3.0
nncf==2.5.0
# deep learning frameworks
tensorflow-macos>=2.5,<=2.12; sys_platform == 'darwin' and platform_machine == 'arm64' # macOS M1 and M2
tensorflow>=2.5,<=2.12; sys_platform == 'darwin' and platform_machine != 'arm64' # macOS x86
tensorflow>=2.5,<=2.12; sys_platform == 'linux' or platform_system == 'Windows'
onnx>=1.11.0
--find-links https://download.pytorch.org/whl/torch_stable.html
torch==1.13.1; sys_platform == 'darwin'
torch==1.13.1+cpu; sys_platform == 'linux' or platform_system == 'Windows'
torchvision==0.14.1; sys_platform == 'darwin'
torchvision==0.14.1+cpu; sys_platform == 'linux' or platform_system == 'Windows'
#paddlepaddle>=2.4.0
transformers>=4.21.1
# others
numpy>=1.21.0
opencv-python
Pillow>=8.3.2
matplotlib>=3.4,<3.5.3
scipy
# The packages below are not directly required. They are dependencies of 
# other dependencies that are pinned to a specific version to avoid
# compatibility issues or vulnerabilities
scikit-image>=0.19.2
setuptools>=56.0.0